Updates Komikcast source
Updates the Komikcast extension to version 0.2.4.

Refactors the tag parsing logic to dynamically fetch genres,
status, types, and sort options from the website, instead of
hardcoding them. This ensures that the extension stays up-to-date
with any changes made to the available filter options on the site.

Updates the URL for fetching tags.